+++ py2app/create_appbundle.py
@@ -1,9 +1,8 @@
+import importlib.resources
 import os
 import plistlib
 import shutil
 import sys
-
-from pkg_resources import resource_filename
 
 import py2app.apptemplate
 from py2app.util import make_exec, makedirs, mergecopy, mergetree, skipscm
@@ -43,11 +42,7 @@ def create_appbundle(
     plistPath = os.path.join(contents, "Info.plist")
     if os.path.exists(plistPath):
         with open(plistPath, "rb") as fp:
-            if hasattr(plistlib, "load"):
-                contents = plistlib.load(fp)
-            else:
-                # 2.7
-                contents = plistlib.readPlist(fp)
+            contents = plistlib.load(fp)
 
             if plist != contents:
                 for d in dirs:
@@ -72,12 +67,13 @@ def create_appbundle(
     print(f"Copy {srcmain!r} -> {destmain!r}")
     copy(srcmain, destmain)
     make_exec(destmain)
-    mergetree(
-        resource_filename(module.__name__, "lib"),
-        resources,
-        condition=condition,
-        copyfn=copy,
-    )
+    with importlib.resources.path(module.__name__, "lib") as p:
+        mergetree(
+            str(p),
+            resources,
+            condition=condition,
+            copyfn=copy,
+        )
     return app, plist