+++ py2app/recipes/black.py
@@ -1,6 +1,4 @@
-import os
-
-from pkg_resources import get_distribution
+from importlib.metadata import packages_distributions
 
 
 def check(cmd, mf):
@@ -8,20 +6,23 @@ def check(cmd, mf):
     if m is None or m.filename is None:
         return None
 
-    egg = get_distribution("black").egg_info
-    top = os.path.join(egg, "top_level.txt")
-
     # These cannot be in zip
-    packages = ["black", "blib2to3"]
+    packages = {"black", "blib2to3"}
 
     # black may include optimized platform specific C extension which has
-    # unusual name, e.g. 610faff656c4cfcbb4a3__mypyc; best to determine it from
-    # the egg-info/top_level.txt
-    with open(top) as f:
-        includes = set(f.read().strip().split("\n"))
-    includes = list(includes.difference(packages))
+    # unusual name, e.g. 610faff656c4cfcbb4a3__mypyc; extract
+    # the name from the list of toplevels.
+    includes = set()
+    for toplevel, dists in packages_distributions():
+        if "black" not in dists:
+            continue
+
+        includes.add(toplevel)
+
+    includes -= packages
 
     # Missed dependency
-    includes.append("pathspec")
+    includes.add("pathspec")
 
-    return {"includes": includes, "packages": packages}
+    # XXX: verify if caller knows how to work with sets
+    return {"includes": list(includes), "packages": list(packages)}